requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.2
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.2.0
autobahn==23.6.2
autogpt==0.0.1.dev0
Automat==22.10.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
boto3==1.34.152
botocore==1.34.152
build==1.2.1
cachetools==5.3.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
chardet==5.1.0
charset-normalizer==3.1.0
chroma-hnswlib==0.7.3
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cohere==5.6.2
coloredlogs==15.0.1
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==41.0.7
cssselect==1.2.0
dataclasses-json==0.6.7
decorator==5.1.1
decouple==0.0.7
Deprecated==1.2.14
deprecation==2.1.0
distro==1.8.0
Django==4.2.10
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-countries==7.5.1
django-ratelimit==4.0.0
django-storages==1.14.2
django-storages-azure==1.6.8
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.3.0
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
elastic-transport==8.11.0
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.111.1
fastapi-cli==0.0.4
fastavro==1.9.5
filelock==3.11.0
flatbuffers==24.3.25
freezegun==1.5.1
frozenlist==1.3.3
fsspec==2024.6.1
google-api-core==2.11.0
google-api-python-client==2.166.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.60.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-geo-type==0.3.12
google-maps-places==0.1.25
google-resumable-media==2.7.1
googleapis-common-protos==1.59.0
gptcache==0.1.44
grpc-google-iam-v1==0.13.1
grpcio==1.65.2
grpcio-status==1.62.2
grpcio-tools==1.62.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.1
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.24.5
humanfriendly==10.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
importlib_metadata==8.0.0
importlib_resources==6.4.0
incremental==22.10.0
iniconfig==2.0.0
instructor==1.3.3
Jinja2==3.1.3
jiter==0.4.2
jmespath==1.0.1
joblib==1.4.2
json_repair==0.25.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
kombu==5.3.5
kubernetes==30.1.0
lancedb==0.5.7
loguru==0.7.0
lxml==4.9.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.21.3
mccabe==0.7.0
mdurl==0.1.2
mem0ai==0.0.9
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.75.0
openpyxl==3.1.2
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.6
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
pgvector==0.2.5
pillow==10.2.0
pip-autoremove==0.10.0
pluggy==1.5.0
portalocker==2.10.1
posthog==3.5.0
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==4.22.3
psycopg2-binary==2.9.9
pulsar-client==3.5.0
py==1.11.0
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.0.1
Pygments==2.18.0
PyJWT==2.8.0
pylance==0.9.18
pyOpenSSL==23.3.0
pyparsing==3.0.9
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pyright==1.1.374
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.2
pytest-django==4.9.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.9
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
qdrant-client==1.10.1
ratelimiter==1.2.0.post0
redis==5.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
retry==0.9.2
rich==13.7.1
rsa==4.9
s3transfer==0.10.2
schema==0.7.7
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.23.1
semver==3.0.2
service-identity==24.1.0
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==2.0.31
sqlparse==0.4.4
starlette==0.37.2
stripe==7.8.1
style==1.1.0
sympy==1.11.1
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.0.0
tqdm==4.65.0
trio==0.26.0
trio-websocket==0.11.1
Twisted==23.10.0
txaio==23.1.1
typer==0.12.3
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
update==0.0.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.22.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.8.2
zipp==3.19.2
zope.interface==6.1
